Split coroutine layout computation to its own file.
`coroutine.rs` was getting too large, and mixing MIR analyses for trait
solving and runtime transformations.
